events: add messageDelete event
Message delete events will be logged in server logs from now.

Note that the content of the deleted message won't be logged or
stored anywhere. Because the content might have some sensitive
data that the user might have sent accidently, and we don't want
to log/store that information.
Logging/storing that is also against Discord's ToS. I had once
asked Dabbit Prime (Discord Staff) and he said it's not allowed
and shouldn't be done.

Refs: https://github.com/TheBastionBot/Bastion/issues/269